requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
pymysql==1.1.0
python-dotenv==1.0.0
WTForms==3.1.1
cryptography==41.0.7
gunicorn==21.2.0
awsgi==0.0.5
email-validator==2.1.0
